refactor(i18n): replace `useServerLocale` hook with `getServerLocale` function and improve typings
- Renamed `useServerLocale` to `getServerLocale` and updated all references accordingly.
- Enhanced typings for `lang` parameters to ensure stricter checks.
- Refactored `interpolateString` to use a shared regex pattern.
- Adjusted state updates for `ui` components (`tracing-beam` and `link-preview`) to include timers and cleanup effects for better performance and reliability.
